autoimpl: separate dereferencable / trait impl usages
